Time: 3 ms (62.03%) | Memory: 23.6 MB (31.07%) - LeetSync

+++ 155-min-stack/min-stack.cpp
@@ -28,11 +28,11 @@ long long mini;
     
     int top() {
         if(st.top()<mini) return mini;
-        return st.top();
+        return (int)(st.top());
     }
     
     int getMin() {
-       return mini;
+       return (int)mini;
     }
 };
